+++ packages/eslint-plugin/src/configs/all.ts
@@ -65,6 +65,8 @@ export = {
     'no-invalid-this': 'off',
     '@typescript-eslint/no-invalid-this': 'error',
     '@typescript-eslint/no-invalid-void-type': 'error',
+    'no-loop-func': 'off',
+    '@typescript-eslint/no-loop-func': 'error',
     'no-loss-of-precision': 'off',
     '@typescript-eslint/no-loss-of-precision': 'error',
     'no-magic-numbers': 'off',
@@ -137,5 +139,7 @@ export = {
     '@typescript-eslint/typedef': 'error',
     '@typescript-eslint/unbound-method': 'error',
     '@typescript-eslint/unified-signatures': 'error',
+    'comma-dangle': 'off',
+    '@typescript-eslint/comma-dangle': 'error',
   },
 };