+++ packages/eslint-plugin/tests/rules/prefer-function-type.test.ts
@@ -1,6 +1,6 @@
 import { AST_NODE_TYPES } from '@typescript-eslint/experimental-utils';
-import rule from '../../src/rules/prefer-function-type';
-import { RuleTester } from '../RuleTester';
+import rule, { phrases } from '../../src/rules/prefer-function-type';
+import { noFormat, RuleTester } from '../RuleTester';
 
 const ruleTester = new RuleTester({
   parserOptions: {
@@ -56,6 +56,9 @@ interface Foo {
         {
           messageId: 'functionTypeOverCallableType',
           type: AST_NODE_TYPES.TSCallSignatureDeclaration,
+          data: {
+            literalOrInterface: phrases[AST_NODE_TYPES.TSInterfaceDeclaration],
+          },
         },
       ],
       output: `
@@ -72,6 +75,9 @@ type Foo = {
         {
           messageId: 'functionTypeOverCallableType',
           type: AST_NODE_TYPES.TSCallSignatureDeclaration,
+          data: {
+            literalOrInterface: phrases[AST_NODE_TYPES.TSTypeLiteral],
+          },
         },
       ],
       output: `
@@ -88,6 +94,9 @@ function foo(bar: { (s: string): number }): number {
         {
           messageId: 'functionTypeOverCallableType',
           type: AST_NODE_TYPES.TSCallSignatureDeclaration,
+          data: {
+            literalOrInterface: phrases[AST_NODE_TYPES.TSTypeLiteral],
+          },
         },
       ],
       output: `
@@ -106,6 +115,9 @@ function foo(bar: { (s: string): number } | undefined): number {
         {
           messageId: 'functionTypeOverCallableType',
           type: AST_NODE_TYPES.TSCallSignatureDeclaration,
+          data: {
+            literalOrInterface: phrases[AST_NODE_TYPES.TSTypeLiteral],
+          },
         },
       ],
       output: `
@@ -124,6 +136,9 @@ interface Foo extends Function {
         {
           messageId: 'functionTypeOverCallableType',
           type: AST_NODE_TYPES.TSCallSignatureDeclaration,
+          data: {
+            literalOrInterface: phrases[AST_NODE_TYPES.TSInterfaceDeclaration],
+          },
         },
       ],
       output: `
@@ -140,10 +155,117 @@ interface Foo<T> {
         {
           messageId: 'functionTypeOverCallableType',
           type: AST_NODE_TYPES.TSCallSignatureDeclaration,
+          data: {
+            literalOrInterface: phrases[AST_NODE_TYPES.TSInterfaceDeclaration],
+          },
         },
       ],
       output: `
 type Foo<T> = (bar: T) => string;
+      `,
+    },
+    {
+      code: `
+interface Foo<T> {
+  (this: T): void;
+}
+      `,
+      errors: [
+        {
+          messageId: 'functionTypeOverCallableType',
+          type: AST_NODE_TYPES.TSCallSignatureDeclaration,
+          data: {
+            literalOrInterface: phrases[AST_NODE_TYPES.TSInterfaceDeclaration],
+          },
+        },
+      ],
+      output: `
+type Foo<T> = (this: T) => void;
+      `,
+    },
+    {
+      code: `
+type Foo<T> = { (this: string): T };
+      `,
+      errors: [
+        {
+          messageId: 'functionTypeOverCallableType',
+          type: AST_NODE_TYPES.TSCallSignatureDeclaration,
+          data: {
+            literalOrInterface: phrases[AST_NODE_TYPES.TSTypeLiteral],
+          },
+        },
+      ],
+      output: `
+type Foo<T> = (this: string) => T;
+      `,
+    },
+    {
+      code: `
+interface Foo {
+  (arg: this): void;
+}
+      `,
+      errors: [
+        {
+          messageId: 'unexpectedThisOnFunctionOnlyInterface',
+          type: AST_NODE_TYPES.TSThisType,
+          data: {
+            interfaceName: 'Foo',
+          },
+        },
+      ],
+    },
+    {
+      code: `
+interface Foo {
+  (arg: number): this | undefined;
+}
+      `,
+      errors: [
+        {
+          messageId: 'unexpectedThisOnFunctionOnlyInterface',
+          type: AST_NODE_TYPES.TSThisType,
+          data: {
+            interfaceName: 'Foo',
+          },
+        },
+      ],
+    },
+    {
+      code: `
+interface Foo {
+  // isn't actually valid ts but want to not give message saying it refers to Foo.
+  (): {
+    a: {
+      nested: this;
+    };
+    between: this;
+    b: {
+      nested: string;
+    };
+  };
+}
+      `,
+      errors: [
+        {
+          messageId: 'functionTypeOverCallableType',
+          type: AST_NODE_TYPES.TSCallSignatureDeclaration,
+          data: {
+            literalOrInterface: phrases[AST_NODE_TYPES.TSInterfaceDeclaration],
+          },
+        },
+      ],
+      output: noFormat`
+type Foo = () => {
+    a: {
+      nested: this;
+    };
+    between: this;
+    b: {
+      nested: string;
+    };
+  };
       `,
     },
   ],