+++ packages/scope-manager/src/lib/es2017.object.ts
@@ -10,6 +10,5 @@ export const es2017_object = {
     isTypeVariable: true,
     isValueVariable: false,
     name: 'ObjectConstructor',
-    writeable: false,
   },
 } as Record<string, ImplicitLibVariableOptions>;