+++ packages/scope-manager/tests/fixtures/jsx/factory/jsxPragma.tsx
@@ -1,0 +1,7 @@
+//// @sourceType = 'module'
+//// @jsxPragma = 'h'
+
+import React from 'react'; // should be unreferenced
+import { h } from 'preact';
+
+<Foo />;