+++ packages/scope-manager/tests/fixtures/type-declaration/conditional4.ts
@@ -1,0 +1,1 @@
+type Test<U> = U extends { [k: string]: infer I } ? I : never;