+++ packages/scope-manager/tests/fixtures/type-declaration/type-parameters/tagged-template.ts
@@ -1,0 +1,6 @@
+type StyledPaymentProps = {
+  isValid: boolean;
+};
+function div<T>(arg: any): void {}
+
+const StyledPayment = div<StyledPaymentProps>``;